requirements.txt
openai
mem0ai
faiss-cpu
dspy
langgraph
langchain
python-dotenv
pandas
numpy
fastapi
uvicorn
pytest